UpdatedUI&LittleBitOfSpriteFormatting
As described on the tin, updated the UI a little more, put in a silly gator easter egg for when I show off the UI, and I imported all available sprites into the sprites folder for quick swap when possible. Also addressed the fact that we were missing a couple sprites for the new tileset, which are being worked on now and will be available by tmw morning. :)